requirements.txt
# Core dependencies for CV Roasting Project

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Gemini API
google-generativeai>=0.3.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0  # For managing API keys

# PDF Processing
PyPDF2>=3.0.0  # Basic PDF text extraction
pdfplumber>=0.10.0  # Advanced PDF parsing (optional but recommended)

# Interactive widgets for Jupyter
ipywidgets>=8.0.0  # For interactive notebook UI
